### 2024-08-22 #7131 @editorialbot
[PRE REVIEW]: StreamGen: a Python framework for generating streams of labeled data
**Submitting author:** @laurenzbeck (<a href="http://orcid.org/0000-0003-1477-1327">Laurenz Farthofer</a>)
**Repository:** https://github.com/Infineon/StreamGen
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
