### 2024-08-27 #7131 @LaurenzBeck
I did a quick search through the [the list of people](https://reviewers.joss.theoj.org/lookup) for potential reviewers and found a few promising reviewer candidates:

- Cedric Kulbach ([kulbachcedric](https://github.com/kulbachcedric))
-  Hoang-Anh Ngo ([hoanganhngo610](https://github.com/hoanganhngo610))
- [mbdemoraes](https://github.com/mbdemoraes)
