### 2024-08-27 c95c7aa
📝 Update README.md
added joss badge
