### 2024-08-27 525d40f
📄 Update paper.bib
added dois for:

- Continual Learning and Catastrophic Forgetting
- Is Class-Incremental Enough for Continual Learning?
