### 2024-08-27 #7131 @LaurenzBeck
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ✅ OK DOIs
>
> - 10.48550/arXiv.2010.15277 is OK
> - 10.1109/TKDE.2018.2876857 is OK
> - 10.48550/arXiv.2106.02585 is OK
> - 10.1109/TSM.2014.2364237 is OK
> - 10.1109/CVPRW53098.2021.00399 is OK
> - 10.48550/arXiv.2102.06253 is OK
> - 10.48550/arXiv.1903.12261 is OK
> - 10.1002/1097-024X(200009)30:11<1203::AID-SPE338>3.0.CO;2-N is OK
>
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: Continual Learning and Catastrophic Forgetting
> - No DOI given, and none found for title: PyTorch: An Imperative Style, High-Performance Dee...
> - No DOI given, and none found for title: TorchVision: PyTorch’s Computer Vision library
> - No DOI given, and none found for title: anytree: Python tree data library
>
> ❌ MISSING DOIs
>
> - 10.3389/frai.2022.829842 may be a valid DOI for title: Is Class-Incremental Enough for Continual Learning...
>
> ❌ INVALID DOIs
>
> - None
> ```

I added DOIs for:
- Is Class-Incremental Enough for Continual Learning?
- Continual Learning and Catastrophic Forgetting
