### 2024-08-27 #7131 @LaurenzBeck
Iztok Fister Jr. ([firefly-cpp](https://github.com/firefly-cpp)) just told me that he would also be willing to review the paper :)
