### 2024-08-28 #7131 @matthewfeickert
Thanks for the reviewer suggestions @LaurenzBeck! These are truly appreciated. I'm traveling and running a conference at the moment so I'm delayed in responding, but before I ask reviewers to spend time on the submission I'm going to ask you to make revisions as the current state of the repository won't pass review given a spot check. Specifically, [there are no automated tests/CI](https://github.com/Infineon/StreamGen/tree/525d40f5f6cfa0c7e2ce5e1eb51a475a1b0f8a65/.github/workflows) which we require or will need a very strong argument for their absence. Please let me know once you've added testing in CI for all versions of Python that you support.

Also, to clarify for the reviewers, you identify the project as a framework, and so given the [upper bounding of all dependencies (except `pytorchcv`)](https://github.com/Infineon/StreamGen/blob/525d40f5f6cfa0c7e2ce5e1eb51a475a1b0f8a65/pyproject.toml#L29-L54) (a library anti-pattern) you are expecting users to engage with the project as an application and _not_ as a library, correct? If so, that's fine, but it should be communicated that this is not meant to be a library, so could you please update references to `streamgen` [being a library](https://github.com/Infineon/StreamGen/blob/525d40f5f6cfa0c7e2ce5e1eb51a475a1b0f8a65/docs/developer_guide.md?plain=1#L67) to application? (_This_ part doesn't need to be done before the review can start though.)

From the rest of my spot check it looks like the rest of the project is in good shape :+1:, though please also read the [Submitting a paper to JOSS](https://joss.readthedocs.io/en/latest/submitting.html#submitting-a-paper-to-joss) docs as well as the [Review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) docs to help you identify and make any other revisions before we progress to adding reviewers.
