### 2024-08-28 #7131 @LaurenzBeck
Thank you @matthewfeickert  for your quick check and feedback 🙏

I will setup the github actions as soon as possible. I am used to other CI/CD platflorms, so I need to do some research/testing first...

The project is indeed a library/framework, since users have to provide their own transformations for their applications.
´streamgen´ just provides the building blocks and tools to do so conveniently :)

The upper bounding using the caret `^` (~ version < next major version) version specification is a default of poetry and ensures working environments without excessive testing when all packages adhere to semantic versioning. I do aggree though, that this might me excessively restrictive and annoying for users, so I will loosen some/all of my bounds. I explicitly added pytorchcv to solve an dependency issue with an optional dependency, I only need when users install `streamgen[cl]`. Since this optional dependency is quite huge (and not so well maintained), I am thinking about copying only the few class definitions I need from those packages.

I could also remove all of the application-specific dependencies, I only need for the examples, and create a seperate application example environment. This would make the project setup and workflows more complex, but users won't be confused by the provided [pip extras](https://github.com/Infineon/StreamGen/blob/525d40f5f6cfa0c7e2ce5e1eb51a475a1b0f8a65/README.md?plain=1#L209). What do you think?

Take your time and enjoy the conference Matthew ;)
