### 2024-08-28 e51cc74
🤖 add python-package workflow
