+++ README.md
@@ -13,7 +13,7 @@ a 🐍 Python framework for generating streams of labelled data
 <p align="center">
     <a href="https://pypi.org/project/streamgen/"><img alt="Static Badge" src="https://img.shields.io/badge/📦_version-1.0.2-blue"></a>
     <a href="https://www.repostatus.org/#active"><img src="https://www.repostatus.org/badges/latest/active.svg" alt="Project Status: Active – The project has reached a stable, usable state and is being actively developed." /></a>
-    <img alt="Static Badge" src="https://img.shields.io/badge/tests-passing-green?logo=pytest">
+    <a href="https://github.com/Infineon/StreamGen/actions/workflows/python-package.yaml"><img alt="🐍 Python package" src="https://github.com/Infineon/StreamGen/actions/workflows/python-package.yaml/badge.svg"></a>
     <img alt="Static Badge" src="https://img.shields.io/badge/Coverage-88%25-yellow?logo=codecov">
 </p>
 
