### 2024-08-28 ece8c14
🤖 modified python-package workflow
