+++ .github/workflows/python-package.yaml
@@ -1,39 +1,33 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
 
-name: 🐍 Python package
+name: 🐍 Python CI
 
-on:
-  push:
-    branches: [ $default-branch ]
-  pull_request:
-    branches: [ $default-branch ]
+on: [push, pull_request]
 
 jobs:
   build:
-
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
       matrix:
         python-version: ["3.10", "3.11", "3.12"]
-
     steps:
-    - uses: actions/checkout@v4
-    - name: 🐍 Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v3
-      with:
-        python-version: ${{ matrix.python-version }}
-    - uses: Gr1N/setup-poetry@v9
-      with:
-        poetry-version: "1.8.3"
-    - uses: actions/cache@v2
-      with:
-        path: ~/.cache/pypoetry/virtualenvs
-        key: ${{ runner.os }}-poetry-${{ hashFiles('poetry.lock') }}
-    - name: 🎶 Install dependencies with poetry
-      run: poetry install --all-extras --sync --compile && poetry shell
-    - name: ⚡ Lint with ruff
-      run: ruff check .
-    - name: ⚗️ Test with pytest
-      run: poe test
+      - uses: actions/checkout@v4
+      - name: 🐍 Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v3
+        with:
+          python-version: ${{ matrix.python-version }}
+      - uses: Gr1N/setup-poetry@v9
+        with:
+          poetry-version: "1.8.3"
+      - uses: actions/cache@v2
+        with:
+          path: ~/.cache/pypoetry/virtualenvs
+          key: ${{ runner.os }}-poetry-${{ hashFiles('poetry.lock') }}
+      - name: 🎶 Install dependencies with poetry
+        run: poetry install --all-extras --sync --compile && poetry shell
+      - name: ⚡ Lint with ruff
+        run: ruff check .
+      - name: ⚗️ Test with pytest
+        run: poe test
