+++ .github/workflows/python-package.yaml
@@ -26,8 +26,8 @@ jobs:
           path: ~/.cache/pypoetry/virtualenvs
           key: ${{ runner.os }}-poetry-${{ hashFiles('poetry.lock') }}
       - name: 🎶 Install dependencies with poetry
-        run: poetry install --all-extras --sync --compile && poetry shell
+        run: poetry install --all-extras --sync --compile
       - name: ⚡ Lint with ruff
-        run: ruff check .
+        run: poetry run ruff check .
       - name: ⚗️ Test with pytest
-        run: poe test
+        run: poetry run poe test
