+++ pyproject.toml
@@ -117,6 +117,7 @@ ignore = [
     "TCH003",
     "TD002",
     "PD901",
+    "EXE001",
     "EXE005",
     "PLR0913",
 ]
