### 2024-08-28 99164d3
🐛 switched to typing_extensions for python 3.10 support
