+++ streamgen/parameter/store.py
@@ -4,9 +4,9 @@ from typing import Any
 
 import pandas as pd
 from beartype import beartype
-from beartype.typing import Self
 from loguru import logger
 from rich.pretty import pretty_repr
+from typing_extensions import Self
 
 from streamgen.parameter import Parameter, ScopedParameterDict
 
