+++ .github/workflows/python-package.yaml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
 
-name: 🐍 Python CI
+name: "🐍 Python CI "
 
 on: [push, pull_request]
 
@@ -13,6 +13,8 @@ jobs:
       matrix:
         python-version: ["3.10", "3.11", "3.12"]
     steps:
+      - name: 🕸️ Install Graphviz
+        run: sudo apt-get install graphviz
       - uses: actions/checkout@v4
       - name: 🐍 Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v3
@@ -21,7 +23,7 @@ jobs:
       - uses: Gr1N/setup-poetry@v9
         with:
           poetry-version: "1.8.3"
-      - uses: actions/cache@v2
+      - uses: actions/cache@v4
         with:
           path: ~/.cache/pypoetry/virtualenvs
           key: ${{ runner.os }}-poetry-${{ hashFiles('poetry.lock') }}
