### 2024-08-28 427d97e
🔥 removed upper dependency constraints
