+++ streamgen/parameter/store.py
@@ -1,12 +1,11 @@
 """🗃️ parameter stores are dictionary-like collections of parameters and schedules."""
 
-from typing import Any
+from typing import Any, Self
 
 import pandas as pd
 from beartype import beartype
 from loguru import logger
 from rich.pretty import pretty_repr
-from typing_extensions import Self
 
 from streamgen.parameter import Parameter, ScopedParameterDict
 
