+++ requirements.txt
@@ -1,96 +1,111 @@
-aiohappyeyeballs==2.3.7 ; python_version >= "3.10" and python_version < "4.0" \
-    --hash=sha256:337ce4dc0e99eb697c3c5a77d6cb3c52925824d9a67ac0dea7c55b8a2d60b222 \
-    --hash=sha256:e794cd29ba6a14078092984e43688212a19081de3a73b6796c2fdeb3706dd6ce
-aiohttp==3.10.4 ; python_version >= "3.10" and python_version < "4.0" \
-    --hash=sha256:03c5a3143d4a82c43a3d82ac77d9cdef527a72f1c04dcca7b14770879f33d196 \
-    --hash=sha256:0fd1f57aac7d01c9c768675d531976d20d5b79d9da67fac87e55d41b4ade05f9 \
-    --hash=sha256:122c26f0976225aba46f381e3cabb5ef89a08af6503fc30493fb732e578cfa55 \
-    --hash=sha256:15b36a644d1f44ea3d94a0bbb71e75d5f394a3135dc388a209466e22b711ce64 \
-    --hash=sha256:1d697023b16c62f9aeb3ffdfb8ec4ac3afd477388993b9164b47dadbd60e7062 \
-    --hash=sha256:2015e4b40bd5dedc8155c2b2d24a2b07963ae02b5772373d0b599a68e38a316b \
-    --hash=sha256:201ddf1471567568be381b6d4701e266a768f7eaa2f99ef753f2c9c5e1e3fb5c \
-    --hash=sha256:23a5f97e7dd22e181967fb6cb6c3b11653b0fdbbc4bb7739d9b6052890ccab96 \
-    --hash=sha256:2ce101c447cf7ba4b6e5ab07bfa2c0da21cbab66922f78a601f0b84fd7710d72 \
-    --hash=sha256:2d64a5a7539320c3cecb4bca093ea825fcc906f8461cf8b42a7bf3c706ce1932 \
-    --hash=sha256:2e685afb0e3b7b861d89cb3690d89eeda221b43095352efddaaa735c6baf87f3 \
-    --hash=sha256:326fb5228aadfc395981d9b336d56a698da335897c4143105c73b583d7500839 \
-    --hash=sha256:347bbdc48411badc24fe3a13565820bc742db3aa2f9127cd5f48c256caf87e29 \
-    --hash=sha256:38bb515f1affc36d3d97b02bf82099925a5785c4a96066ff4400a83ad09d3d5d \
-    --hash=sha256:394ddf9d216cf0bd429b223239a0ab628f01a7a1799c93ce4685eedcdd51b9bc \
-    --hash=sha256:438c6e1492d060b21285f4b6675b941cf96dd9ef3dfdd59940561029b82e3e1f \
-    --hash=sha256:45bb655cb8b3a61e19977183a4e0962051ae90f6d46588ed4addb8232128141c \
-    --hash=sha256:46cc9069da466652bb7b8b3fac1f8ce2e12a9dc0fb11551faa420c4cdbc60abf \
-    --hash=sha256:4ad284cee0fdcdc0216346b849fd53d201b510aff3c48aa3622daec9ada4bf80 \
-    --hash=sha256:4b34e5086e1ead3baa740e32adf35cc5e42338e44c4b07f7b62b41ca6d6a5bfd \
-    --hash=sha256:50ac670f3fc13ce95e4d6d5a299db9288cc84c663aa630142444ef504756fcf7 \
-    --hash=sha256:5115490112f39f16ae87c1b34dff3e2c95306cf456b1d2af5974c4ac7d2d1ec7 \
-    --hash=sha256:58df59234be7d7e80548b9482ebfeafdda21948c25cb2873c7f23870c8053dfe \
[…]
