+++ .github/workflows/python-package.yaml
@@ -11,7 +11,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.10", "3.11", "3.12"]
+        python-version: ["3.11", "3.12"]
     steps:
       - name: 🕸️ Install Graphviz
         run: sudo apt-get install graphviz
