### 2024-08-28 973b0a1
🔥 removed upper dependency constraints when possible
