### 2024-08-29 b1f56de
🤖 added CI; 🐍 dropped python 3.10 support; 📦 removed upper bounds for dependencies
