+++ .github/workflows/python-package.yaml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
 
-name: "🐍 Python CI "
+name: CI
 
 on: [push, pull_request]
 
