+++ docs/paper.bib
@@ -50,6 +50,7 @@
 @misc{cossu_is_2021,
 	title = {Is {Class}-{Incremental} {Enough} for {Continual} {Learning}?},
 	url = {http://arxiv.org/abs/2112.02925},
+	doi = {10.3389/frai.2022.829842},
 	abstract = {The ability of a model to learn continually can be empirically assessed in different continual learning scenarios. Each scenario defines the constraints and the opportunities of the learning environment. Here, we challenge the current trend in the continual learning literature to experiment mainly on class-incremental scenarios, where classes present in one experience are never revisited. We posit that an excessive focus on this setting may be limiting for future research on continual learning, since class-incremental scenarios artificially exacerbate catastrophic forgetting, at the expense of other important objectives like forward transfer and computational efficiency. In many real-world environments, in fact, repetition of previously encountered concepts occurs naturally and contributes to softening the disruption of previous knowledge. We advocate for a more in-depth study of alternative continual learning scenarios, in which repetition is integrated by design in the stream of incoming information. Starting from already existing proposals, we describe the advantages such class-incremental with repetition scenarios could offer for a more comprehensive assessment of continual learning models.},
 	urldate = {2023-08-29},
 	publisher = {arXiv},
@@ -83,6 +84,7 @@
 @misc{van_de_ven_continual_2024,
 	title = {Continual {Learning} and {Catastrophic} {Forgetting}},
 	url = {http://arxiv.org/abs/2403.05175},
+	doi = {10.48550/arXiv.2403.05175},
 	abstract = {This book chapter delves into the dynamics of continual learning, which is the process of incrementally learning from a non-stationary stream of data. Although continual learning is a natural skill for the human brain, it is very challenging for artificial neural networks. An important reason is that, when learning something new, these networks tend to quickly and drastically forget what they had learned before, a phenomenon known as catastrophic forgetting. Especially in the last decade, continual learning has become an extensively studied topic in deep learning. This book chapter reviews the insights that this field has generated.},
 	language = {en},
 	urldate = {2024-03-15},
