### 2024-08-29 #7131 @LaurenzBeck
Hi @matthewfeickert 👋

- I added a github workflow to perform linting and testing for all supported python versions on each push and merge request
- I removed most upper version bounds to help with the dependency resolution of users. Thank you for the hint ❤️
- While setting up the CI, I noticed, that I forgot to mention graphviz as a manual/system dependency and added the information in the [installation instructions](https://github.com/Infineon/StreamGen/tree/main?tab=readme-ov-file#-installation)
- I released all minor changes as a patch  (1.0.3)
