### 2024-09-26 Infineon/StreamGen#5 @firefly-cpp
Tested OS
Is there any evidence that the software works on most major operating systems? Maybe you can add a line stating where this has already been tested in a similar style as outlined here: https://github.com/firefly-cpp/NiaAML

This issue is a part of the review process: openjournals/joss-reviews#7206
