### 2024-09-26 Infineon/StreamGen#5 @LaurenzBeck
I don't think there should be surprises since most dependencies are pure Python ones 🤔

I already tested on windows locally and the CI runs on an ubuntu machine.
I will have a look if I can include different operating systems in the CI test matrix.

Thanks for the hint Iztok 🙏
