### 2024-09-30 4331182
🤖 added windows and macOs to the CI
