+++ .github/workflows/python-package.yaml
@@ -7,11 +7,12 @@ on: [push, pull_request]
 
 jobs:
   build:
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
       matrix:
         python-version: ["3.11", "3.12"]
+        os: ["ubuntu-latest", "windows-latest", "macos-latest"]
     steps:
       - name: 🕸️ Install Graphviz
         run: sudo apt-get install graphviz
