### 2024-09-30 f1eec5b
☕ added CI runners for windows and macOs
