+++ .github/workflows/python-package-windows.yaml
@@ -1,21 +1,20 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
 
-name: CI
+name: Windows CI
 
 on: [push, pull_request]
 
 jobs:
   build:
-    runs-on: ${{ matrix.os }}
+    runs-on: windows-latest
     strategy:
       fail-fast: false
       matrix:
         python-version: ["3.11", "3.12"]
-        os: ["ubuntu-latest", "windows-latest", "macos-latest"]
     steps:
       - name: 🕸️ Install Graphviz
-        run: sudo apt-get install graphviz
+        run: choco install graphviz
       - uses: actions/checkout@v4
       - name: 🐍 Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v3
