### 2024-09-30 Infineon/StreamGen#5 @LaurenzBeck
I added runners for windows and macOs :)

No surprises there 🎉
