### 2024-10-04 Infineon/StreamGen#5 @hoanganhngo610
Same results here. I tried installing on both Windows and MacOS, both running perfectly fine!
