### 2024-10-07 Infineon/StreamGen#6 @firefly-cpp
Paper comments
Hi @LaurenzBeck,

I am posting some relevant comments you may consider when revising your software paper, which is currently under review in JOSS.

According to https://github.com/pytorch/vision, the author is outlined as "TorchVision maintainers and contributors." You should also also make it consistent in your paper.

You may also devote some sentences to theoretical background to continual learning in the beginning when you start introducing terms. Also, add a couple of significant references, review papers, etc.

Define acronyms on the site when they appear for the first time.

You may also need to include a more detailed description of proposed software architecture. This aspect should be considered since it also helps future generations of users quickly learn, modify, and adopt it. Also, why not present one Python snippet in your paper? See one example of a paper here: https://joss.theoj.org/papers/10.21105/joss.06748.pdf

You should also explicitly mention similar or (competitive) frameworks that are close to your proposed framework, along with the pros and cons (one table?).

This issue is a part of the review process: https://github.com/openjournals/joss-reviews/issues/7206
