### 2024-10-08 5f3b426
worked on paper references
