+++ docs/paper.bib
@@ -155,12 +155,11 @@
 	file = {Paszke et al. - PyTorch An Imperative Style, High-Performance Dee.pdf:C\:\\Users\\Hundgeburth\\Zotero\\storage\\XZP46DKD\\Paszke et al. - PyTorch An Imperative Style, High-Performance Dee.pdf:application/pdf},
 }
 
-@misc{maintainers_torchvision_2016,
-	title = {{TorchVision}: {PyTorch}'s {Computer} {Vision} library},
-	url = {https://github.com/pytorch/vision},
-	author = {maintainers, TorchVision and {contributors}},
-	month = nov,
-	year = {2016},
+@software{torchvision2016,
+    title = {{{TorchVision}}: {{PyTorch}}'s {{Computer Vision}} Library},
+    author = {TorchVision maintainers and contributors},
+    date = {2016-11},
+    url = {https://github.com/pytorch/vision}
 }
 
 @misc{c0fec0de_anytree_2016,
