### 2024-10-08 Infineon/StreamGen#6 @LaurenzBeck
morning Iztok ☕👋

Thank you for also checking the paper contents so thoroghly 🙏

> According to https://github.com/pytorch/vision, the author is outlined as "TorchVision maintainers and contributors." You should also also make it consistent in your paper.

I corrected the bib entry of torchvision, Zotero messed this on up...

> You may also devote some sentences to theoretical background to continual learning in the beginning when you start introducing terms. Also, add a couple of significant references, review papers, etc.

This one was a tricky one for me, since Continual Learning research was my main motivation of creating the package, but it is really just one application StreamGen. I feel like I already spend a lot of paper space (briefly) introducing it in the Statement of Need section (lines 44 - 78).

> You may also need to include a more detailed description of proposed software architecture.

I quickly went over https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain gain to make sure this is not a strict requirement. It would be a nice section in a full length software paper, but with 200 words above the upper recommended bound, I feel like this would stretch this JOSS paper too much. I explained architectural choices and details in a lot of detail in the [user guide of the documentation](https://infineon.github.io/StreamGen/examples/time%20series%20classification/01-static-distributions/).

> Also, why not present one Python snippet in your paper?

Same reason here, the documentation contains many examples, and I wanted to stay within the word count bounds.
Unfortunately, a representative and educational code example would probably easily fit 1 - 2 pages :/

> You should also explicitly mention similar or (competitive) frameworks that are close to your proposed framework, along with the pros and cons (one table?).

I do mention other similar edavours in the statement of need section:

- stream construction by splitting with cons (line 48 - 59)
- datasets generated by means of data augmentation (line 70)
- simulation based appoaches (lines 71 - 74)

I also mention how Streamgen extends these ideas (lines 75 - 77).

Given that there are no direct alternatives with concrete criteria to compare and distuinguish StreamGen, I would favour the free text format over a table.

---

I will leave this issue open for you to respond and for @hoanganhngo610 to also use it for his comments ;)

best wishes,
Laurenz
