### 2024-10-08 Infineon/StreamGen#6 @hoanganhngo610
Thank you so much @firefly-cpp for your first comments into the paper. @LaurenzBeck, my first comment would also be regarding the references, since there are several papers in your bibiography cited as a preprint from arXiV, while in fact they aren't anymore. These include:

- *Is Class-Incremental Enough for Continual Learning?*, published in Frontiers in Artificial Intelligence;
- *Benchmarking Neural Network Robutness to Common Corruptions and Pertubations*, published in ICLR 2019 (available on [OpenReview](https://openreview.net/forum?id=HJz6tiCqYm));
- *A Procedural World Generation Framework for Systematic Evaluation of Continual Learning*, published in NeurIPS 2021 Datasets and Benchmarks Track (available on [OpenReview](https://openreview.net/forum?id=LlCQWh8-pwK));
- *Class-Incremental Learning: Survey and Performance Evaluation on Image Classification*, published on IEEE Transactions on Patter Analysis and Machine Intelligence (available on [IEEE](https://doi.org/10.1109/TPAMI.2022.3213473))
