### 2024-10-08 d0b9179
📝 updated arxiv references
