+++ docs/paper.bib
@@ -1,18 +1,19 @@
 
-@misc{masana_class-incremental_2022,
-	title = {Class-incremental learning: survey and performance evaluation on image classification},
-	shorttitle = {Class-incremental learning},
-	url = {http://arxiv.org/abs/2010.15277},
-	doi = {10.48550/arXiv.2010.15277},
-	abstract = {For future learning systems, incremental learning is desirable because it allows for: efficient resource usage by eliminating the need to retrain from scratch at the arrival of new data; reduced memory usage by preventing or limiting the amount of data required to be stored -- also important when privacy limitations are imposed; and learning that more closely resembles human learning. The main challenge for incremental learning is catastrophic forgetting, which refers to the precipitous drop in performance on previously learned tasks after learning a new one. Incremental learning of deep neural networks has seen explosive growth in recent years. Initial work focused on task-incremental learning, where a task-ID is provided at inference time. Recently, we have seen a shift towards class-incremental learning where the learner must discriminate at inference time between all classes seen in previous tasks without recourse to a task-ID. In this paper, we provide a complete survey of existing class-incremental learning methods for image classification, and in particular, we perform an extensive experimental evaluation on thirteen class-incremental methods. We consider several new experimental scenarios, including a comparison of class-incremental methods on multiple large-scale image classification datasets, an investigation into small and large domain shifts, and a comparison of various network architectures.},
-	urldate = {2023-05-09},
-	publisher = {arXiv},
-	author = {Masana, Marc and Liu, Xialei and Twardowski, Bartlomiej and Menta, Mikel and Bagdanov, Andrew D. and van de Weijer, Joost},
-	month = oct,
-	year = {2022},
-	note = {arXiv:2010.15277 [cs]},
-	keywords = {Computer Science - Machine Learning, Computer Science - Computer Vision and Pattern Recognition},
-	file = {arXiv Fulltext PDF:C\:\\Users\\Hundgeburth\\Zotero\\storage\\WZJFLLNS\\Masana et al. - 2022 - Class-incremental learning survey and performance.pdf:application/pdf;arXiv.org Snapshot:C\:\\Users\\Hundgeburth\\Zotero\\storage\\RRJFHSWE\\2010.html:text/html},
+@article{masana_class-incremental_2022,
+	title = {Class-{{Incremental Learning}}: {{Survey}} and {{Performance Evaluation}} on {{Image Classification}}},
+	shorttitle = {Class-{{Incremental Learning}}},
+	author = {Masana, Marc and Liu, Xialei and Twardowski, Bartłomiej and Menta, Mikel and Bagdanov, Andrew D. and family=Weijer, given=Joost, prefix=van de, useprefix=true},
+	date = {2023-05},
+	journaltitle = {IEEE Transactions on Pattern Analysis and Machine Intelligence},
+	volume = {45},
+	number = {5},
+	pages = {5513--5533},
+	issn = {1939-3539},
+	doi = {10.1109/TPAMI.2022.3213473},
+	url = {https://ieeexplore.ieee.org/document/9915459},
+	urldate = {2024-10-08},
[…]
