### 2024-10-08 Infineon/StreamGen#6 @LaurenzBeck
Thank you a lot @hoanganhngo610 for checking the citations 🙏

Did you check them manually? Can you recommend me a workflow/tool on how to do it more efficiently? (I mostly have pre-print references in my database 😅)
I quickly checked, Zotero is not able to do it automatically.

Anyways, I updated the 4 references you mentioned.
