### 2024-10-11 #7206 @matthewfeickert
(I was offline for a few days, so responding now.)

Thanks @firefly-cpp and @hoanganhngo610. The timelines sound good, but as there might be some back and forth needed in discussions I'll set reminders for 2 weeks from now. :+1:
