### 2024-10-16 Infineon/StreamGen#6 @hoanganhngo610
@LaurenzBeck Thank you so much for your response, and for the prompt action to modify the references. In fact, I have to check all of them manually, and I don't really think there would be any workflow/tool that can do it efficiently. Whenever I have to write papers, I have to double-check to make sure that the cited preprints actually did not appear somewhere else.

I will continue my review for the paper and/or discussion on the review of @firefly-cpp within the next comments.
