### 2024-10-16 Infineon/StreamGen#6 @hoanganhngo610
My first follow-up comment based on the suggestion from @firefly-cpp would be:

> This one was a tricky one for me, since Continual Learning research was my main motivation of creating the package, but it is really just one application of StreamGen. I feel like I already spend a lot of paper space (briefly) introducing it in the Statement of Need section (lines 44 - 78).
The main focus of StreamGen would be to create a framework to generate data streams, and it has multiple applications apart from Continual Learning itself. However, the author can add 1-2 more sentences about the application and constrains of CL within the first paragraph (lines 44-47), for example each sample is only used once, the number of tasks to be learned are not pre-defined, or the accomodation of new information should not occur without catastrophic forgetting and inteference. Only mentioning that CL "... extends the paradigm to work in dynamic (...) enviornments ..." and "... acquire and preserve knowledge continually" seems not enough.
