### 2024-10-17 Infineon/StreamGen#6 @LaurenzBeck
I guess I am already to deep into the topic and asuming too much.
Thank you for the clarification, I will add and re-write some parts today 😊
