### 2024-10-17 e1690c3
iterated on the paper
