+++ docs/paper.md
@@ -12,7 +12,7 @@ authors:
     orcid: 0000-0003-1477-1327
     affiliation: "1, 2" # (Multiple affiliations must be quoted)
 affiliations:
-  - name: Kai - Kompetenzzentrum Automobil- und Industrieelektronik GmbH, Austria
+  - name: KAI - Kompetenzzentrum Automobil- und Industrieelektronik GmbH, Austria
     index: 1
   - name: Institute of Computer Graphics and Vision, Graz University of Technology, Austria
     index: 2
@@ -20,15 +20,15 @@ date: 19 August 2024
 bibliography: paper.bib
 ---
 
-![A tree of sampling functions and transformations as a new data structure and framework for synthetic data generation. Samples are generated by traversing the tree from the root to the leaves. Each path through the tree represents its own class-conditional distribution. The branching points represent categorical distributions which determine the path to take for a sample during the tree traversal. By changing the parameters of the transformations over time, such trees can represent evolving distribution suitable to generate data streams (see \autoref{fig:parameter_schedule}).\label{fig:sampling_tree}](images/sampling_tree.png){ width=95% }
+![A tree of sampling functions and transformations as a new data structure and framework for synthetic data generation. Samples are generated by traversing the tree from the root to the leaves. Each path through the tree represents its own class-conditional distribution. The branching points represent categorical distributions which determine the path to take for a sample during the tree traversal. By changing the parameters of the transformations over time, such trees can represent evolving distributions suitable to generate data streams (see \autoref{fig:parameter_schedule}).\label{fig:sampling_tree}](images/sampling_tree.png){ width=95% }
 
 # Summary
 
 StreamGen is a framework for generating streams of labeled, synthetic data from trees composed of sampling functions and transformation monoids (see \autoref{fig:sampling_tree}).
 
-Due to the expensive nature of the labelling process, researchers and machine learning practitioners often rely on existing datasets and stochastic data augmentation pipelines like `torchvision.transforms.Compose` objects [TorchVision @torchvision2016]. While such methods and datasets are enough to study learning from static domains, emerging research fields like continual learning study learning on long streams of data, representing evolving experiences. StreamGen addresses this need by giving researchers a tool to model time-dependent, diverse class-conditional distributions.
+Due to the expensive nature of the labelling process, researchers and machine learning practitioners often rely on existing datasets and stochastic data augmentation pipelines like `torchvision.transforms.Compose` objects [TorchVision @torchvision2016]. While such methods and datasets are appropriate to study learning from static domains, emerging research fields like continual learning study learning on long streams of data, representing evolving experiences. StreamGen addresses this need by giving researchers a tool to model time-dependent, diverse class-conditional distributions.
 
-Such distributions can be represented through the use of a [tree](https://en.wikipedia.org/wiki/Tree_(data_structure)) data structure (or other more general linked structures like directed acyclic graphs (DAG)) to store sampling functions and transformations. Samples are generated by traversing the tree from the root to the leaves. Each branching point represents a categorical distribution which determines the path to take for a sample during the tree traversal.
+Such distributions can be represented through the use of a [tree](https://en.wikipedia.org/wiki/Tree_(data_structure)) data structure (or other more general linked structures like directed acyclic graphs) to store sampling functions and transformations. Samples are generated by traversing the tree from the root to the leaves. Each branching point represents a categorical distribution which determines the path to take for a sample during the tree traversal. This information can be utilized for automating the annotation process.
 
[…]
