### 2024-10-17 Infineon/StreamGen#6 @LaurenzBeck
Done: https://github.com/Infineon/StreamGen/actions/runs/11379909229

I iterated on some explanations and managed to keep the current page layout  😁
