### 2024-10-25 Infineon/StreamGen#6 @LaurenzBeck
are you satisfied with the results @firefly-cpp and @hoanganhngo610 ?
If so, I will close the Issue :)
