### 2024-10-28 Infineon/StreamGen#6 @hoanganhngo610
@LaurenzBeck Sorry I did not have much time to look into your updates within the paper. I will have a look at it by Friday and inform you by Friday if there is anything else needed.
