### 2024-11-07 Infineon/StreamGen#6 @LaurenzBeck
Hi @hoanganhngo610 👋

Did you had time to check the paper?

Iztok gave his green lights for the JOSS release this week :D
