### 2024-11-08 Infineon/StreamGen#6 @hoanganhngo610
@LaurenzBeck I have just had a look at your changes within the paper, and that is sufficient for the moment. Thank you very much for the great work, and also a very well-presented paper as well!
