### 2024-11-11 #7206 @matthewfeickert
@LaurenzBeck I'll take a few days this week to go through the editorial checklist and then once I've done that I'll prompt you for any minor revisions to the text and then give you guidance on what next steps to take to get everything ready for publication.
