### 2024-11-11 30d7f2d
🤖 updated CI badge in README.md
