### 2024-12-02 Infineon/StreamGen#7 @matthewfeickert
[JOSS Review] Paper questions and editorial nitpicks
This issue is with regards to the JOSS review happening at https://github.com/openjournals/joss-reviews/issues/7206.

Overall, I want to say that you did a great job of capturing the spirit of what a JOSS paper is supposed to be, so thank you for that! Along these lines though, do you feel that Figure 2

https://github.com/Infineon/StreamGen/blob/30d7f2deeede8612e9e7a6ea996eacc6ca510cb6/docs/paper.md?plain=1#L52

is strictly necessary? Or is Figure 1 sufficient and then the two places where Figure 2 is referred to

https://github.com/Infineon/StreamGen/blob/30d7f2deeede8612e9e7a6ea996eacc6ca510cb6/docs/paper.md?plain=1#L23

https://github.com/Infineon/StreamGen/blob/30d7f2deeede8612e9e7a6ea996eacc6ca510cb6/docs/paper.md?plain=1#L33

could just refer to the documentation? If you feel Figure 2 is necessary / critically relevant that's totally fine, but it might be useful to reduce the figure caption in that case to a more succinct summary.

Now getting into the truly nitpicky bits, there are a few places where a hyphen is used where an em-dash should be used instead.

* `represent the *domain* - the data distribution common to all classes.` in

https://github.com/Infineon/StreamGen/blob/30d7f2deeede8612e9e7a6ea996eacc6ca510cb6/docs/paper.md?plain=1#L52C243-L52C312

* `from past experiences - a phenomenon referred to as` in

https://github.com/Infineon/StreamGen/blob/30d7f2deeede8612e9e7a6ea996eacc6ca510cb6/docs/paper.md?plain=1#L57C1-L58C1

In these situations the hyphen (-) should be replaced with an em-dash (&mdash;). As the papers are being powered by Markdown and not pure LaTeX we can't reach for the easy LaTeX `---` but you can use `&mdash;` which will insert an em-dash (note the `;` at the end which is actually relevant).

* As https://github.com/c0fec0de/anytree doesn't have a `CITATION.cff` there's no easy citation there for you to take (that's fine) but for

https://github.com/Infineon/StreamGen/blob/30d7f2deeede8612e9e7a6ea996eacc6ca510cb6/docs/paper.md?plain=1#L38

is there a particular release version that is relevant to the mention in the paper that could be cited? Or just the entire project?

* Can you try to escape with quotes

https://github.com/Infineon/StreamGen/blob/30d7f2deeede8612e9e7a6ea996eacc6ca510cb6/docs/paper.bib#L154

so that the authors show up as "TorchVision maintainers and contributors" in the references instead of "maintainers, T., & contributors"?

* I see in the review that `editorialbot` is giving a skipped DOI for

given that

https://github.com/Infineon/StreamGen/blob/30d7f2deeede8612e9e7a6ea996eacc6ca510cb6/docs/paper.bib#L144-L150

is in the `paper.bib` but apparently not used. If PyTorch doesn't need to be directly cited can you remove that from the `paper.bib` to avoid these catches?

Other than that I think this is great. :+1: Feel free to ask questions if you have any!
