### 2024-12-02 #7206 @matthewfeickert
@LaurenzBeck Sorry for being dreadfully slow, but I've got just a few questions and nitpicks over on https://github.com/Infineon/StreamGen/issues/7 and then we can publish this. :+1:
