### 2024-12-02 Infineon/StreamGen#7 @matthewfeickert
By the way, if it is at all useful to preview changes to the paper locally rather than having `editorialbot` rebuild it (which to be clear, is fine) you can use Docker to build the paper on your local machine using these instructions: https://joss.readthedocs.io/en/latest/paper.html#docker
