### 2024-12-02 Infineon/StreamGen#7 @LaurenzBeck
> Overall, I want to say that you did a great job of capturing the spirit of what a JOSS paper is supposed to be

thank you @matthewfeickert 🙏, this means a lot to me.

I will go through your comments now.
