### 2024-12-02 Infineon/StreamGen#7 @LaurenzBeck
> In these situations the hyphen (-) should be replaced with an em-dash (—).

Nice catch, I just learned about this last week in a scientific writing seminar. I will change the hyphens 💪
