### 2024-12-02 Infineon/StreamGen#7 @LaurenzBeck
> is in the paper.bib but apparently not used. If PyTorch doesn't need to be directly cited can you remove that from the paper.bib to avoid these catches?

removed 😊
