### 2024-12-02 Infineon/StreamGen#7 @LaurenzBeck
> Can you try to escape with quotes so that the authors show up as "TorchVision maintainers and contributors" in the references instead of "maintainers, T., & contributors"?

Quotes resulted in "maintainers, TorhVision, and contributors" (or similar).
Surrounding the field with double { (author = {{{TorchVision maintainers and contributors}}},) did the trick 😊
