### 2024-12-02 Infineon/StreamGen#7 @LaurenzBeck
I have a formatting question regarding the list of affiliations at the beginning of the paper:

Is there a way to include a linebreak between the two entries?
I would prefer something like:

Laurenz A. Farthofer 1,2
1) KAI - Kompetenzzentrum Automobil- und Industrieelektronik GmbH, Austria
2) Institute of Computer Graphics and Vision, Graz University of Technology, Austria
