### 2024-12-02 Infineon/StreamGen#7 @LaurenzBeck
Ok, I checked anytree and I do not depend on the major changes introduced in the current 2.x versions.
Therefore, I would like to cite the project as a whole.

How do I need to adapt the citation entry?
