### 2024-12-02 Infineon/StreamGen#7 @LaurenzBeck
I had a look at the caption of Figure 2:
I don't need any additional space, so I personally don't mind the longer caption. (I always prefer self-contained figures)
If you would still prefer a shorter one, we could remove the detailed explanation after the first sentence in the caption.

I tried and it looks very acceptable with the statement of need section starting on the second page. I could try 1 or two tricks/rephrasings to save ~3  lines, which would result in the last page only containing references. (Now it also contains the last parahraph from the acknowledgements...
What do you think @matthewfeickert ?
