### 2024-12-03 Infineon/StreamGen#7 @matthewfeickert
> Is there a way to include a linebreak between the two entries?

Good question. I'm not sure, but I'll ask and if support for this gets added post publication we have the ability to rerender a paper.

> Ok, I checked anytree and I do not depend on the major changes introduced in the current 2.x versions. Therefore, I would like to cite the project as a whole.

In your `pyproject.toml` you require the latest release to date of the project

https://github.com/Infineon/StreamGen/blob/30d7f2deeede8612e9e7a6ea996eacc6ca510cb6/pyproject.toml#L32

You can select whatever version requirements you'd like, but if you decided to be more library-like then this would ideally be an empirically tested lower bound. `anytree` doesn't make citation super easy given that they don't provide any citation guidance or long term archives with DOIs, but from [_Software citation principles_](https://peerj.com/articles/cs-86/) (2016) one of the motivations of software citation is specificity

> ... Software identification should be as specific as necessary, such as using version numbers, revision numbers, or variants such as platforms.

So I'd be inclined to cite what you require, which regardless of if you're actually using the features, in StreamGen's case is `v2.12.1`.

> I don't need any additional space, so I personally don't mind the longer caption. (I always prefer self-contained figures). If you would still prefer a shorter one, we could remove the detailed explanation after the first sentence in the caption.

I'd suggest keeping it as short as possible to explain **only** the _need for the software_. I am in agreement with you that figures should give the necessary context themselves, but at seven sentences long this is a full paragraph on technical information that the documentation should cover in full. If the documentation doesn't then I'd suggest moving the detailed explanation here to the documentation and then simply giving a 1-2 sentence summary and citing the docs.
