### 2024-12-03 Infineon/StreamGen#7 @LaurenzBeck
good, I will check the lowest version for which my test suit passes and then update both the dependency in `pyproject.toml`, as well as the reference in `paper.bib` :)

I will also shorten the figure caption and check if the docs contain the detailed explanations.
Should I explicitly refer to the docs for more details in the figure caption, or is this obvious?

Thank you for raising the question in inara 💪
