+++ docs/paper.bib
@@ -141,28 +141,22 @@
 	langid = {english},
 }
 
-@article{paszke_pytorch_nodate,
-	title = {{PyTorch}: {An} {Imperative} {Style}, {High}-{Performance} {Deep} {Learning} {Library}},
-	abstract = {Deep learning frameworks have often focused on either usability or speed, but not both. PyTorch is a machine learning library that shows that these two goals are in fact compatible: it provides an imperative and Pythonic programming style that supports code as a model, makes debugging easy and is consistent with other popular scientiﬁc computing libraries, while remaining efﬁcient and supporting hardware accelerators such as GPUs.},
-	language = {en},
-	author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
-	file = {Paszke et al. - PyTorch An Imperative Style, High-Performance Dee.pdf:C\:\\Users\\Hundgeburth\\Zotero\\storage\\XZP46DKD\\Paszke et al. - PyTorch An Imperative Style, High-Performance Dee.pdf:application/pdf},
-}
-
 @software{torchvision2016,
     title = {{{TorchVision}}: {{PyTorch}}'s {{Computer Vision}} Library},
-    author = {TorchVision maintainers and contributors},
+    author = {{{TorchVision maintainers and contributors}}},
     date = {2016-11},
-    url = {https://github.com/pytorch/vision}
+    url = {https://github.com/pytorch/vision},
+	version = {0.20.1},
 }
 
-@misc{c0fec0de_anytree_2016,
+@software{c0fec0de_anytree_2016,
 	title = {anytree: {Python} tree data library},
 	url = {https://github.com/c0fec0de/anytree},
 	publisher = {GitHub},
 	author = {{c0fec0de}},
 	year = {2016},
 	note = {Publication Title: GitHub repository},
+	version = {2.12.1},
 }
 
 @article{gansner_open_1997,
