+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "streamgen"
-version = "1.0.3"
+version = "1.0.4"
 description = "🌌 a framework for generating streams of labeled data."
 authors = ["Laurenz A. Farthofer <laurenz@hey.com>"]
 packages = [{ include = "streamgen" }]
@@ -29,27 +29,27 @@ build_paper = "docker run --rm --volume $PWD/docs:/data --env JOURNAL=joss openj
 [tool.poetry.dependencies]
 python = "^3.11"
 # 🌳 Graphs and Trees
-anytree = ">=2.12.1"  # Apache license 2.0
+anytree = ">=2.7.0"
 graphviz = ">=0.20.3"
 # 📄 Data handling
-numpy = ">=1.26.4" # custom license -> seems permissive
-pandas = ">=2.2.1" # BSD 3-clause license
+numpy = ">=1.26.4"
+pandas = ">=2.2.1"
 # 🖼️ Visualizations
-matplotlib = ">=3.8.3" # custom license -> seems permissive
-seaborn = ">=0.13.2"   # multiple licenses -> seems permissive (like MIT) if licenses are included
+matplotlib = ">=3.8.3"
+seaborn = ">=0.13.2"
 # ✍️ Typing
-beartype = "<0.18" # MIT license
+beartype = "<0.18"
 # 🖨️ Logging
-loguru = ">=0.7.2" # MIT license
-rich = ">=13.7.1"  # MIT license
+loguru = ">=0.7.2"
+rich = ">=13.7.1"
 # 🔁 Continual Learning
-#avalanche-lib = { version = "^0.5.0", optional = true } # MIT license
-continuum = { version = ">=1.2.7", optional = true }     # MIT license
+#avalanche-lib = { version = "^0.5.0", optional = true }
+continuum = { version = ">=1.2.7", optional = true }
 # version pins required for an import error caused by avalanche
-torch = { version = ">=2.3", optional = true }        # MIT license
-pytorchcv = { version = ">=0.0.67", optional = true } # MIT license
+torch = { version = ">=2.3", optional = true }
+pytorchcv = { version = ">=0.0.67", optional = true }
 # 👾 Example/Application Dependencies
-polars = { version = ">=0.20.13", optional = true }    # MIT license
+polars = { version = ">=0.20.13", optional = true }
 ipympl = { version = ">=0.9.4", optional = true }
 scikit-image = { version = ">=0.24", optional = true }
 
@@ -79,7 +79,7 @@ jupyter = ">=1.0.0"
 ipywidgets = ">=8.1.2"
 nbconvert = ">=7.16.2"
 portray = ">=1.8.0"
-itables = ">=2.0.0" # MIT license
+itables = ">=2.0.0"
 qpsolvers = { extras = [
     "open-source-solvers",
 ], version = ">=4.3.2" } # explicit install required to solve plotting issue in pytest
