### 2024-12-03 9d63df3
⚡ worked on new ruff lints
