+++ .pre-commit-config.yaml
@@ -7,12 +7,12 @@ repos:
         stages: [post-checkout, post-merge]
         always_run: true
   - repo: https://github.com/python-poetry/poetry-plugin-export.git
-    rev: 1.7.1
+    rev: 1.8.0
     hooks:
       - id: poetry-export
         args: [-f, requirements.txt, -o, requirements.txt, --with, dev, --all-extras]
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.5.0
+    rev: v5.0.0
     hooks:
       - id: check-added-large-files
         args: [--maxkb=10240]
@@ -30,7 +30,7 @@ repos:
         exclude: .ipynb
         args: [--autofix]
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.3.4
+    rev: v0.8.1
     hooks:
       - id: ruff
         args: [--fix, --exit-non-zero-on-fix, --config=pyproject.toml]
