### 2024-12-03 Infineon/StreamGen#7 @LaurenzBeck
ok, I checked and unfortunately, I depend on anytree being >=2.7 since I make use of the `UniqueDotExporter`, which was introduced then. I lowered the dependency constrain to >=2.7.0

I also added package versions to anytree and torchvision, which look great in the references.

I also shortened the figure caption and removed some whitespace/new-lines in the "future work" section to have the last page only show the references.

Finally, I did a patch release of streamgen to make the lower dependency constraints applicable for end users :)

How long do we wait for an answer from inara#95 @matthewfeickert ?
