### 2024-12-03 #7206 @matthewfeickert
@LaurenzBeck all that's left to do now before recommending publication is to ensure that there is a long term public archive of the code that was reviewed. In this case the last code release was [`StreamGen` `v1.4.0`](https://github.com/Infineon/StreamGen/releases/tag/1.4.0), which you just made today, but for simplicity of what follows you might want to consider making a `v1.4.1` release (as I don't think you have a public archive). But read on and you can decide what to do and we'll go from there. :+1:

We'd suggest depositing the code either with [Zenodo](https://zenodo.org/) or with [figshare](https://figshare.com/) to get an archive with a DOI. If you use Zenodo there is [an (optional) GitHub integration](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) that can create a Zenodo archive for you anytime you make a GitHub release of your code.

[![Zenodo_DOI_guide](https://user-images.githubusercontent.com/5142394/234173148-2226325d-0a42-4526-b546-75b469d26c4c.png)](https://zenodo.org/account/settings/github/)

Once you make a release of `StreamGen` with a DOI for the archive, and have verified all related author tasks listed in https://github.com/openjournals/joss-reviews/issues/7206#issuecomment-2467417899, please just report them here and I'll have @editorialbot update the version listed in the review and add the archive.

Let me know if you have any questions. :+1:
