### 2024-12-04 ce71b81
📝 Zenodo archive release for the JOSS publication
