+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "streamgen"
-version = "1.0.4"
+version = "1.0.5"
 description = "🌌 a framework for generating streams of labeled data."
 authors = ["Laurenz A. Farthofer <laurenz@hey.com>"]
 packages = [{ include = "streamgen" }]
