### 2024-12-04 768dccd
📘 create CITATION.cff
