+++ CITATION.cff
@@ -1,0 +1,9 @@
+cff-version: 1.1.0
+message: "🤚 if you use this software, please cite it as below:"
+authors:
+- family-names: Farthofer
+  given-names: Laurenz
+orcid: https://orcid.org/0000-0003-1477-1327
+title: StreamGen: a Python framework for generating streams of labeled data
+version: 1.0.5
+date-released: 2024-12-04
