### 2024-12-04 316de3c
📘 update CITATION.cff
