### 2024-12-04 2445468
📝 added Zenodo DOI badge to `README.md`
