### 2024-12-04 #7206 @LaurenzBeck
@matthewfeickert , I now have a DOI for the 1.0.5 JOSS-specific and archived release:
https://doi.org/10.5281/zenodo.14273611
