### 2024-12-04 #7206 @matthewfeickert
Oh, I guess they both do as they both have archive names of "Infineon-StreamGen-316de3c". I'm not sure how two got generated with the same name, but if that didn't happen as some accidental trigger on your side I'd report that to the Zenodo team to avoid confusion. As they're the same commit I'll take the first one (https://doi.org/10.5281/zenodo.14273611) and use that.
