### 2024-12-05 #7206 @matthewfeickert
> I then deleted the release and the tag and released the same version again with the cff file, which resulted in the DOI I sent you (https://doi.org/10.5281/zenodo.14273611).

@LaurenzBeck Okay, as long as we have the one that matches what you're intendeding as the `v1.0.5` release (which we do :+1:) that's all that matters.

> Do we need to do sth about the second entry?

For publication in JOSS, no, but it might be generally useful to contact Zenodo support (https://zenodo.org/support) to remove the other `v1.0.5` archive (https://doi.org/10.5281/zenodo.14273653) as that one is showing up as being newer than the one you've selected, and so might be confusing to others.
