### 2024-12-05 #7206 @crvernon
:wave: @LaurenzBeck - I just need you to make the following changes before I move to accept this for publication:

In the paper:
- [ ] LINE 11:  There should be a semicolon after the first "TorchVision" in "...(TorchVision TorchVision main..."
- [ ] LINES 88-89:  Should "...with special characteristic represent other..." be "...with special characteristics representing..."?

Let me know when these have been addressed and we will move forward.  Thanks!
