+++ docs/paper.md
@@ -10,7 +10,7 @@ tags:
 authors:
   - name: Laurenz A. Farthofer
     orcid: 0000-0003-1477-1327
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+    affiliation: "1, 2"
 affiliations:
   - name: KAI - Kompetenzzentrum Automobil- und Industrieelektronik GmbH, Austria
     index: 1
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 StreamGen is a framework for generating streams of labeled, synthetic data from trees composed of sampling functions and transformation monoids (see \autoref{fig:sampling_tree}).
 
-Due to the expensive nature of the labelling process, researchers and machine learning practitioners often rely on existing datasets and stochastic data augmentation pipelines like `torchvision.transforms.Compose` objects [TorchVision @torchvision2016]. While such methods and datasets are appropriate to study learning from static domains, emerging research fields like continual learning study learning on long streams of data, representing evolving experiences. StreamGen addresses this need by giving researchers a tool to model time-dependent, diverse class-conditional distributions.
+Due to the expensive nature of the labelling process, researchers and machine learning practitioners often rely on existing datasets and stochastic data augmentation pipelines like `torchvision.transforms.Compose` objects [@torchvision2016]. While such methods and datasets are appropriate to study learning from static domains, emerging research fields like continual learning study learning on long streams of data, representing evolving experiences. StreamGen addresses this need by giving researchers a tool to model time-dependent, diverse class-conditional distributions.
 
 Such distributions can be represented through the use of a [tree](https://en.wikipedia.org/wiki/Tree_(data_structure)) data structure (or other more general linked structures like directed acyclic graphs) to store sampling functions and transformations. Samples are generated by traversing the tree from the root to the leaves. Each branching point represents a categorical distribution which determines the path to take for a sample during the tree traversal. This information can be utilized for automating the annotation process.
 
@@ -77,7 +77,7 @@ StreamGen builds on these ideas and provides researchers with a general and intu
 
 # Future work
 
-The generation of multi-label samples requires loops and cycles for a compact and convenient representation. Such scenarios are still representable with tree data structures by unrolling these cycles through many redundant paths and transformations. A representation using less restricted types of graphs presents an interesting future extension to the framework. StreamGen already defines protocols and base classes to include different sampler concepts in the future. More declarative ways to build schedules and distributions with special characteristic represent other promising extensions.
+The generation of multi-label samples requires loops and cycles for a compact and convenient representation. Such scenarios are still representable with tree data structures by unrolling these cycles through many redundant paths and transformations. A representation using less restricted types of graphs presents an interesting future extension to the framework. StreamGen already defines protocols and base classes to include different sampler concepts in the future. More declarative ways to build schedules and distributions represent other promising extensions.
 
 # Acknowledgements
 
