### 2024-12-06 #7206 @LaurenzBeck
Hi @crvernon 👋

Thank you for your feedback 🙏❤️
After so many iterations, it is so easy to oversee such things.

I removed the redundant "TorchVision" from the reference in line 11, which stems from an earlier version where the `authors` property was not rendered properly (surrounding it with double curly-braces in the bib entry had solved the problem).
I also removed the unnecessary and vague "... with special characteristics ...", which makes it clear, that "represent" refers so the "declarative ways" and not the "special characteristics" 😊

The changes are already pushed to the repository.
