### 2024-12-06 62a95fa
👋 update CITATION.cff
